2025/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.7.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
